Refactor pinot-csv out
