+++ pom.xml
@@ -54,6 +54,7 @@
     <module>pinot-hadoop-filesystem</module>
     <module>pinot-azure-filesystem</module>
     <module>pinot-record-readers/pinot-avro</module>
+    <module>pinot-record-readers/pinot-csv</module>
     <module>pinot-record-readers/pinot-orc</module>
     <module>pinot-record-readers/pinot-parquet</module>
     <module>pinot-connectors</module>
